Version: 0.2.3a (task 4.20 — Skill «Экономия денег» / Money Saving: -1% к тратам — done).
Новый Gym-навык, снижающий стоимость денежных трат. Применяется к Gym
training + Shop purchases. НЕ применяется к Work salary (доход), Bank
operations (exploit-prone), Adventure (там нет money cost).

Формула — линейная cost * (1 - skill/100) (выбрана пользователем, несмотря
на «риск 0$» при skill=100; альтернативы diminishing returns / cap 50%
отвергнуты ради простоты). При skill > 100 — clamp до 0.

Реализация:
- state.GymSkills.money_saving: int = 0 — round-trip flat-key.
- bonus.apply_money_saving(cost: float, state) -> float = round(
    max(0.0, cost * (1 - skill/100)), 2). Единая точка округления, макс 2
  знака после запятой.
- actions.try_spend(money: float = 0.0) — сигнатура поднята с int.
  state.money уже float с 0.2.2.
- gym.py — 4 точки cost['money'] обёрнуты: format_lvl_up_info /
  get_lvl_up_info (display :,.2f), check_requirements (compare +
  сообщение), start_skill_training (try_spend).
- shop.py — 4 точки покупок (чизбургер / кофе / 3 кед):
  _buy_item(cost: float), display цен через apply_money_saving с :,.2f.
- web/main.py — _build_gym_skills и _validate_and_apply_training учитывают
  скидку. Сообщение о нехватке :,.2f.
- gym._SKILL_DESCRIPTIONS['money_saving'] — title «Экономия денег».
- gym_menu skill_options пункт '13', потом reorder на '9' рядом с
  money-related навыками (banking_interest_rate, loan_capacity,
  loan_interest_reduction, inspiration сдвинуты на одну позицию вниз).
- web._GYM_SKILL_DISPLAY['money_saving'] — icon 🏷, available=True.
  Порядок entries определяет порядок рендера — money_saving тоже на 9-м.

12 новых тестов:
- 8 в test_bonus.py (no skill / linear / fractional 750→697.50 / round до
  2 знаков / skill=100→0 / skill=150 clamp / cost=0 / тип float)
- 4 в test_bank.py (skill в _SKILL_DESCRIPTIONS / _GYM_SKILL_DISPLAY,
  try_spend принимает float, _buy_item с дисконтом).

Документация: TASKS.md (4.20 → done; добавлена follow-up 4.20.1 «Аудит
округления цен / денег» по запросу пользователя), changelog.txt (0.2.3a),
CLAUDE.md (счётчик навыков 12 → 13; раздел level.py / bonus.py с описанием
apply_money_saving + 8 точек применения + edge cases).

All 522 tests pass (510 + 12), mypy 0 issues. Version 0.2.3 → 0.2.3a.

+++ actions.py
@@ -15,7 +15,7 @@ from typing import Optional
 from state import GameState
 
 
-def try_spend(state: GameState, steps: int = 0, energy: int = 0, money: int = 0) -> bool:
+def try_spend(state: GameState, steps: int = 0, energy: int = 0, money: float = 0.0) -> bool:
     """Атомарно проверяет ресурсы и списывает их.
 
     Возвращает True, если все ресурсы покрывают запрошенную стоимость,
